re-establish full object meta-data
For some reason object metadata gets whipped out when a field is touched on that object. When sent to the cumulusci team they narrowed down the issue to the fact that we have enhanced lookup enabled on our objects. It appears to be a Salesforce CLI issue. I have not experienced this lately after running sfdx update in my command line. I suspect, but can't confirm, that this issue might be fixed.