reestablishing object metadata
Metecho tends to wipe out object metadata if you could any fields on the object. I believe this is a Salesforce CLI error when enhanced lookup is on the object. I have updated my Salesforce CLI and the issue stopped appearing for me. I think Metecho may be a little behind in CLI updates.