Description, Help Text and Field Labe Update
These are updates from the October Sprint that never got into Master. Most are Description and Help Text updates, but there is also a Field Label change on Appointment Type as "UG Event Appointment Type" is still there from the original. This is one that could have other impacts when packaging, so worth noting next time we package.