#431 Resolution - Require Event on Manually Created Reg Records
Update Registration Page Layout so Event is required if creating an internally new registration.  Created a Quick Action on instance that has a default Event and Instance so only Registrant information is needed (Name, Email, Phone).  remove "UG" from a related list label as that was UST Specific in past. Added the Quick Action to the Instance Layout.